fix: support HA
