+++ sqlx-core/src/postgres/options/mod.rs
@@ -8,9 +8,11 @@ mod connect;
 mod parse;
 mod pgpass;
 mod ssl_mode;
+mod target_session_attrs;
+use crate::error::Error;
 use crate::{connection::LogSettings, net::CertificateInput};
 pub use ssl_mode::PgSslMode;
-use crate::error::Error;
+pub use target_session_attrs::TargetSessionAttrs;
 
 /// Options and flags which can be used to configure a PostgreSQL connection.
 ///
@@ -93,32 +95,6 @@ pub struct PgConnectOptions {
     pub(crate) extra_float_digits: Option<Cow<'static, str>>,
     pub(crate) options: Option<String>,
     pub(crate) target_session_attrs: TargetSessionAttrs,
-}
-
-#[derive(Debug, Clone, Default, Eq, PartialEq)]
-pub enum TargetSessionAttrs {
-    /// No special properties are required.
-    #[default]
-    Any,
-    /// The session must allow writes.
-    ReadWrite,
-}
-
-impl FromStr for TargetSessionAttrs {
-    type Err = Error;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Ok(match &*s.to_ascii_lowercase() {
-            "any" => TargetSessionAttrs::Any,
-            "read-write" => TargetSessionAttrs::ReadWrite,
-
-            _ => {
-                return Err(Error::Configuration(
-                    format!("unknown value {:?} for `target_session_attrs`", s).into(),
-                ));
-            }
-        })
-    }
 }
 
 impl Default for PgConnectOptions {
